backend/requirements.txt
# CrewAI Framework
crewai>=0.86.0
crewai-tools>=0.17.0

# Google Gemini SDK (for TTS and Embeddings)
google-genai>=1.0.0
google-generativeai>=0.8.0

# Audio processing
pydub>=0.25.1

# Environment variables
python-dotenv>=1.0.0

# Web Frontend
flask>=3.0.0
flask-cors>=4.0.0
